Fix keyboard shortcut agents ref and extract NumericInput
