+++ App.tsx
@@ -471,10 +471,6 @@ const App: React.FC = () => {
         latestHandleRun.current = handleRun;
     }, [handleRun]);
 
-    useEffect(() => {
-        agentsRef.current = agents;
-    }, [agents]);
-
 
     useEffect(() => {
         const handler = (e: KeyboardEvent) => {
@@ -585,7 +581,8 @@ const App: React.FC = () => {
 
     useEffect(() => {
         isHistoryViewRef.current = displayData.isHistoryView;
-    }, [displayData.isHistoryView]);
+        agentsRef.current = displayData.agents;
+    }, [displayData]);
 
     const handleSaveAll = async () => {
         const dataToSave = displayData;
